+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+0.21.0 - (2022-04-17)
+------------------
+
+* Fixed `--log` argument in `music_into_network.py` commands.
+* Fixed ability to set output side length in `music_into_network.py` commands.
+* Added alternate frames -> video writer functionality that invokes ffmpeg directly, for better
+control of output quality. "High RQuality" settings are optimized for uploading big videos to 
+YouTube. Default settings still use MP4V encoder.
+* Note: This version that produced the production _Won Pound_ music videos.
+
+  
 0.20.0 - (2022-03-29)
 ------------------